requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
bidict==0.23.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
markdown2==2.5.2
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
nicegui==2.8.1
numpy==2.2.2
ocpp==0.20.0
opentelemetry-api==1.29.0
opentelemetry-exporter-zipkin==1.29.0
opentelemetry-exporter-zipkin-json==1.29.0
opentelemetry-exporter-zipkin-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
packaging==24.2
propcache==0.2.1
protobuf==3.20.3
pscript==0.7.7
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-json-logger==3.2.1
python-multipart==0.0.20
python-socketio==5.12.1
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.0.4
websockets==14.2
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0